Added comments + use images from HERE
